+++ RELEASENOTES.md
@@ -84,6 +84,8 @@
     `C.MSG_SET_OUTPUT_BUFFER_RENDERER`.
   * Use `VideoDecoderRenderer` as an implementation of
     `VideoDecoderOutputBufferRenderer`, instead of `VideoDecoderSurfaceView`.
+* Add automatic audio becoming noisy handling to `SimpleExoPlayer`,
+  available through `SimpleExoPlayer.setHandleAudioBecomingNoisy`.
 * Add `Timeline.Window.isLive` to indicate that a window is a live stream
   ([#2668](https://github.com/google/ExoPlayer/issues/2668) and
   [#5973](https://github.com/google/ExoPlayer/issues/5973)).